pyproject.toml
[project]
name = "spriteforge"
version = "0.1.0"
description = "An AI-powered spritesheet generator for 2D games."
readme = "README.md"
requires-python = ">=3.12.3,<3.13"
authors = [{ name = "Francesco Sodano" }]
license = { text = "MIT" }
dependencies = [
    "pillow>=12.1.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "azure-ai-projects>=1.0.0b7",
    "azure-identity>=1.21.0",
    "aiohttp>=3.10.0",
]

[project.optional-dependencies]
azure = [
    "azure-ai-projects>=1.0.0b7",
    "azure-identity>=1.21.0",
    "aiohttp>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black==26.1.0",
    "mypy==1.19.1",
    "pytest==9.0.2",
    "pytest-asyncio>=0.23",
    "python-dotenv>=1.2.1",
    "types-pillow>=10.2.0.20240822",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that require live Azure AI Foundry access (DefaultAzureCredential + AZURE_AI_PROJECT_ENDPOINT)",
]

[tool.pytest-asyncio]
mode = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
target-version = ["py312"]
